+++ services/director-v2/src/simcore_service_director_v2/modules/dynamic_sidecar/scheduler/_core/_observer.py
@@ -8,8 +8,10 @@ from math import floor
 from fastapi import FastAPI
 from servicelib.error_codes import create_error_code
 
-from .....core.dynamic_services import DynamicServicesSettings
-from .....core.dynamic_services.scheduler import DynamicServicesSchedulerSettings
+from .....core.dynamic_services_settings import DynamicServicesSettings
+from .....core.dynamic_services_settings.scheduler import (
+    DynamicServicesSchedulerSettings,
+)
 from .....core.dynamic_sidecar_settings import DynamicSidecarSettings
 from .....models.dynamic_services_scheduler import (
     DynamicSidecarStatus,
